+++ packages/vui-wysiwyg/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vui-wysiwyg
 
+## 4.0.76
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.16.76
+
 ## 4.0.75
 
 ### Patch Changes